pyproject.toml
[tool.poetry]
name = "realtime_app"
version = "0.1.0"
description = ""
authors = ["Kirill <kir.pystin@gmail.com>"]

packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "3.8"
aiohttp = "^3.8.1"
aiomisc = "^16.0.8"
aiohttp-middlewares = "^1.2.1"
ConfigArgParse = "^1.5.3"
aio-pika = "^7.2.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run_server = "app.__main__:run"
run_generator = "app.workers.generator:run"